Classifier: add @store/WorkflowStore directory Move WorkflowStore files into their own directory.
