+++ packages/lib-classifier/src/store/WorkflowStore/WorkflowStore.js
@@ -1,7 +1,7 @@
 import { autorun } from 'mobx'
 import { addDisposer, getRoot, isValidReference, types } from 'mobx-state-tree'
-import ResourceStore from './ResourceStore'
-import Workflow from './Workflow'
+import ResourceStore from '../ResourceStore'
+import Workflow from '../Workflow'
 import queryString from 'query-string'
 
 const WorkflowStore = types
